+++ RedditPodcastPoster.Common/PodcastServices/IFlushable.cs
@@ -1,0 +1,6 @@
+﻿namespace RedditPodcastPoster.Common.PodcastServices;
+
+public interface IFlushable
+{
+    void Flush();
+}
