+++ RedditPodcastPoster.Common/PodcastServices/Spotify/ICachedSpotifyClient.cs
@@ -1,10 +1,8 @@
-﻿using Microsoft.Extensions.Logging;
-using RedditPodcastPoster.Common.PodcastServices.Apple;
-using SpotifyAPI.Web;
+﻿using SpotifyAPI.Web;
 
 namespace RedditPodcastPoster.Common.PodcastServices.Spotify;
 
-public interface ICachedSpotifyClient: IFlushable
+public interface ICachedSpotifyClient : IFlushable
 {
     CachedSpotifyClient.CachedEpisodesClient Episodes { get; }
     CachedSpotifyClient.CachedShowsClient Shows { get; }
@@ -22,34 +20,4 @@ public interface ICachedSpotifyClient: IFlushable
         IndexingContext indexingContext,
         IPaginator? paginator = null,
         CancellationToken cancel = default);
-
-}
-
-public interface IFlushable
-{
-    void Flush();
-}
-
-public class CacheFlusher : IFlushable
-{
-    private readonly ICachedSpotifyClient _cachedSpotifyClient;
-    private readonly ICachedApplePodcastService _cachedApplePodcastService;
-    private readonly ILogger<CacheFlusher> _logger;
-
-    public CacheFlusher(
-        ICachedSpotifyClient cachedSpotifyClient,
-        ICachedApplePodcastService cachedApplePodcastService,   
-        ILogger<CacheFlusher> logger)
-    {
-        _cachedSpotifyClient = cachedSpotifyClient;
-        _cachedApplePodcastService = cachedApplePodcastService;
-        _logger = logger;
-    }
-
-    public void Flush()
-    {
-        _logger.LogInformation($"{nameof(Flush)}");
-        _cachedApplePodcastService.Flush();
-        _cachedSpotifyClient.Flush();
-    }
 }
